# src/modules/feedback/dto/add-feedback.dto.ts
import { IsNotEmpty, IsString } from 'class-validator';
import {ApiProperty} from "@nestjs/swagger";

export class AddFeedbackDto {
  @IsString()
  @IsNotEmpty()
  @ApiProperty()
  feedback: string;
}